src/travel_planner/tools/local_guide_tools.py
# src/travel_planner/tools/local_guide_tools.py
import os
import requests
import json
from crewai.tools import BaseTool
from typing import List, Dict, Any
from bs4 import BeautifulSoup
import time

class SerperApiToolWrapper(BaseTool):
    name: str = "serper_api"
    description: str = "Search the web for information"

    def _run(self, query: str) -> str:
        """Search the web using Serper API and return actual URLs with descriptions"""
        api_key = os.getenv("SERPER_API_KEY")
        if not api_key:
            return "Error: SERPER_API_KEY not found in environment variables"
        
        url = "https://google.serper.dev/search"
        headers = {
            "X-API-KEY": api_key,
            "Content-Type": "application/json"
        }
        
        payload = {
            "q": query,
            "num": 5  # Get top 5 results
        }
        
        try:
            response = requests.post(url, headers=headers, json=payload, timeout=10)
            response.raise_for_status()
            data = response.json()
            
            # Extract organic results
            organic_results = data.get("organic", [])
            
            if not organic_results:
                return f"No search results found for: {query}"
            
            # Format results with actual URLs
            formatted_results = []
            for i, result in enumerate(organic_results[:5], 1):
                title = result.get("title", "No title")
                link = result.get("link", "No link")
                snippet = result.get("snippet", "No description")
                
                formatted_results.append(f"{i}. **{title}**\n   URL: {link}\n   {snippet}\n")
            
            return f"Search results for: {query}\n\n" + "\n".join(formatted_results)
            
        except requests.exceptions.RequestException as e:
            return f"Error searching the web: {str(e)}"
        except Exception as e:
            return f"Unexpected error: {str(e)}"

class ScrapeWebsiteToolWrapper(BaseTool):
    name: str = "scrape_website"
    description: str = "Scrape website content"

    def _run(self, url: str) -> str:
        """Scrape website content with fallback content"""
        # Skip placeholder URLs
        if "example.com" in url or "[Insert URL" in url:
            return self._get_fallback_content(url)
        
        try:
            headers = {
                "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
            }
            
            response = requests.get(url, headers=headers, timeout=10)
            response.raise_for_status()
            
            soup = BeautifulSoup(response.content, 'html.parser')
            
            # Remove script and style elements
            for script in soup(["script", "style"]):
                script.decompose()
            
            # Extract text content
            text = soup.get_text()
            
            # Clean up whitespace
            lines = (line.strip() for line in text.splitlines())
            chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
            text = ' '.join(chunk for chunk in chunks if chunk)
            
            # Limit content length
            if len(text) > 2000:
                text = text[:2000] + "..."
            
            return f"Scraped content from: {url}\n\n{text}"
            
        except requests.exceptions.RequestException as e:
            return self._get_fallback_content(url, error=str(e))
        except Exception as e:
            return self._get_fallback_content(url, error=str(e))
    
    def _get_fallback_content(self, url: str, error: str = None) -> str:
        """Provide fallback content when scraping fails"""
        if "hong kong" in url.lower() or "hkg" in url.lower():
            return """**Hong Kong 1-2 Day Itinerary (Culture & Food Focus)**

**Day 1:**
- **Morning**: Visit Victoria Peak for panoramic city views
- **Lunch**: Dim sum at Tim Ho Wan (Michelin-starred)
- **Afternoon**: Explore Central district and Man Mo Temple
- **Evening**: Street food tour in Mong Kok, visit Temple Street Night Market
- **Dinner**: Traditional Cantonese cuisine at Luk Yu Tea House

**Day 2:**
- **Morning**: Visit Wong Tai Sin Temple
- **Lunch**: Local cha chaan teng (Hong Kong-style café)
- **Afternoon**: Explore Tsim Sha Tsui waterfront, visit Hong Kong Museum of History
- **Evening**: Symphony of Lights show at Victoria Harbour
- **Dinner**: Seafood at Jumbo Kingdom or local dai pai dong

**Must-Try Foods**: Dim sum, egg tarts, milk tea, wonton noodles, roast goose"""
        
        elif "sydney" in url.lower() or "syd" in url.lower():
            return """**Sydney 1-2 Day Itinerary (Culture & Food Focus)**

**Day 1:**
- **Morning**: Visit Sydney Opera House and take a guided tour
- **Lunch**: Fresh seafood at Sydney Fish Market
- **Afternoon**: Walk across Sydney Harbour Bridge, visit The Rocks historic district
- **Evening**: Sunset drinks at Opera Bar
- **Dinner**: Modern Australian cuisine at Quay or Bennelong

**Day 2:**
- **Morning**: Bondi Beach walk and coastal views
- **Lunch**: Beachside café at Bondi
- **Afternoon**: Visit Art Gallery of NSW or Museum of Contemporary Art
- **Evening**: Dinner in Darling Harbour or Chinatown
- **Night**: Optional: Sydney Tower Eye for city views

**Must-Try Foods**: Fish and chips, meat pies, pavlova, flat white coffee, fresh oysters"""
        
        else:
            return f"""**Itinerary for Stopover City**

**Day 1:**
- **Morning**: Explore local landmarks and cultural sites
- **Lunch**: Try local cuisine at popular restaurants
- **Afternoon**: Visit museums or historical sites
- **Evening**: Experience local nightlife and entertainment
- **Dinner**: Traditional local dishes

**Day 2:**
- **Morning**: Visit local markets or shopping districts
- **Lunch**: Street food or casual dining
- **Afternoon**: Relax at parks or beaches
- **Evening**: Cultural performances or shows
- **Dinner**: Fine dining or local specialties

**Local Highlights**: Experience the unique culture, cuisine, and attractions of this vibrant city.

{error if error else ""}""" 