requirements.txt
crewai>=0.103.0
crewai-tools>=0.0.1
streamlit>=1.20.0
requests>=2.28.0
geopy>=2.3.0
python-dotenv>=0.21.0
torch>=2.0.0
amadeus>=8.0.0
google-cloud-aiplatform>=1.44.0
beautifulsoup4>=4.12.0